refactor utils funcs
