+++ src/components/load-more-btn.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import {createElement} from '../utils/render.js';
 
 const createLoadMoreButtonTemplate = () => {
   return (
