+++ src/components/site-menu.js
@@ -1,5 +1,5 @@
 import {filterCount} from '../mock/filter.js';
-import {createElement} from "../utils.js";
+import {createElement} from '../utils/render.js';
 
 const createSiteMenuTemplate = (films) => {
   return (
