+++ src/components/user-rate.js
@@ -1,5 +1,5 @@
 import {filterCount} from '../mock/filter.js';
-import {createElement} from "../utils.js";
+import {createElement} from '../utils/render.js';
 
 const createUserRateTemplate = (films) => {
   return (
