+++ src/mock/filter.js
@@ -1,5 +1,3 @@
-const filterCount = (films, parametr) => {
+export const filterCount = (films, parametr) => {
   return films.filter((it) => it[parametr] === true).length;
 };
-
-export {filterCount};
